Various fixes for the firebird adapter
Even with these changes, Sequel's support on Firebird is minimal at
best.  It seems fairly easy to get the connection into a state
where subsequent queries fail, even with the changes I added to
attempt to prevent this.

I tried getting the firebird adapter to work with the standard fb
gem instead of wishdev's fork, but I got request synchronization
errors for most queries.

Here's a list of changes made:

* Add disconnect support for certain error conditions.
* Remove the downcase_names setting, which shouldn't be needed.
* Fetch rows as arrays instead of hashes, as the resulting hashes
  still need to be processed, and using arrays is faster.
* Remove Database#primary_key_for_table, which shouldn't have been
  public.
* Make Database#primary_key only accept a single argument, and fix
  it so it works correctly.  This allows insert to correctly use
  the primary key value.
* Clear cached primary keys when dropping tables.
* Support Database#views
* Remove unused Dataset#hash_row method.