Support limiting of one_to_many associations when eager loading
This is fairly simple support that just implements an array slice
in ruby after all associated objects have been eagerly loaded.
It gives correct results, but doesn't offer a performance
improvement as it still loads all rows from the database.  It's
designed to be extensible with future strategies that modify the
eager loading query to limit the rows.