Add support for an :eager_limit_strategy => :window_function value for limiting using window functions
This works for all association types that do eager limiting.  It's
the default type used by datasets that support window functions,
assuming you are using :eager_limit_strategy=>true to enable the
support (except for one_to_one on PostgreSQL, which uses
:distinct_on).

Basically, the row_number window function is added to the query in
a subselect, and the main query does a filter on the row number
to remove rows outside the range of row numbers you are looking
for.  The row_number window function uses a partition so there is
a separate series for each of the main objects, which is how you
can get eager loading to work correctly.

This can be much faster than the default :ruby method for limiting
eagerly loaded associations, if your limit is small compared to
the number of rows actually associated.