Make Dataset#select_append check for DB2 syntax issue with SELECT *, column not being supported
This adds Dataset#supports_select_all_and_column? for checking
for support, which is currently only false on DB2.  If the value
is false and no columns are selected, it automatically selects
all columns from all FROM and JOIN tables to emulate *.  This
cleans up the DB2 offset emulation and makes the new
:window_function eager limit strategy work on DB2.