Work around issues with using LIMIT/OFFSET in IN subqueries on MySQL, MSSQL, and DB2
MySQL just has a flat out bug where you can't use a LIMIT in the IN
subquery, but you can run the query as a nested subquery without
any problems.

MSSQL and DB2 emulate offset support using ROW_NUMBER, which modifies
the SELECTed columns to add the ROW_NUMBER.  This breaks use with IN
as the subquery should only be returning a single column.  So this
runs a nested subquery that only selects the first column.

To DRY up some code, move emulated offset support into a separate
module that both the shared mssql and db2 adapters include.