Add support for an :eager_limit_strategy => :correlated_subquery value for limiting using correlated subqueries
This will never be used by default as it can generate very slow
queries if not used wisely.  But for databases that don't support
window functions, this may be the best way to eagerly load limited
associations.

Unforatunately, due to MySQL bugs, the :correlated_subquery
strategy doesn't work on MySQL.  MySQL doesn't allow a LIMIT
directly in an IN subselect, and when you work around that using a
nested subquery, it appears MySQL doesn't handle the correlation
properly.

This also doesn't work correctly on DB2, probably because DB2's
scoping only looks in the directly enclosing scope and not all
enclosing scopes.

This also contains another spec for the window_function case
in many_through_many, some spec description fixes in
many_through_many, as well as a fix if the core extensions
are not enabled.