Add Database#copy_table for PostgreSQL COPY support when using the postgres adapter with pg
If you want to get CSV formatted data, this can be much faster than
getting the results normally and formatting them in ruby.

PostgreSQL actually has three separate COPY syntaxes, one before
7.3, one from 7.3 to 8.4, and one starting in 9.0.  While 9.0+
supports earlier syntaxes, I choose to support only the 9.0+
syntax.  You can provide the COPY command as a string if you
want to run on versions before 9.0.

If the COPY results are not fully retrieved, the connection is
left in an unstable state, so I choose to just raise a
DatabaseDisconnectError in that case.  Another option would be
to attempt to fully retrieve results, but that has more corner
cases.