Add support for WITH in INSERT/UPDATE/DELETE and RETURNING in UPDATE/DELETE on Postgresql 9.1
This is a fairly large commit that changes some of the core parts
of Sequel.  It adds a Dataset#returning method that sets the columns
to return, and changes Dataset #insert, #update, and #delete to
yield the returned rows to a block if #returning is used.  The rows
are yielded as plain hashes, as it uses #fetch_rows instead of #each.
I may add the ability to use #each in the future, but I'll wait to
see if anyone needs that.  To work correctly with the master/slave
support, when #returning is used with #insert, #update, or #delete,
the :default server is used if no server has been set.

The only database that I know of that supports these features is
PostgreSQL 9.1, but the changes were too invasive to make to just
the shared postgres adapter.  Hopefully, future versions of other
database libraries will support the syntax as well.

The changes to the shared postgres adapter are significant.
Dataset#insert was modified signficantly to use the new returning
feature.  That require changes to the bound variable/prepared
statement support in the shared postgres adapter.

There are a few backwards incompatible changes to the PostgreSQL
and Firebird adapters.  The previously public dataset method
something different. The private #insert_returning_pk_sql and
unlikely user code would be calling these methods anyway.  I haven't
tested the Firebird support, but I will try to soon.

Another backwards incompatible change is th Dataset::WITH_SUPPORTED
constant has been removed.

Dataset#supports_cte? now takes an optional type argument, which
defaults to :select for backwards compatibility.  It can be now be
used for checking for WITH support in INSERT/UPDATE/DELETE.
Dataset#supports_returning? has been added.  It's type argument is
required, and it can be used for checking whether RETURNING is
supported in INSERT/UPDATE/DELETE.