Support the :limit association option when using eager_graph
Currently, this only does an array slice in ruby, it doesn't add
support for various eager limit strategies.  Supporting such
strategies in eager graph is possible, but I'm not sure it's wise
as it would have to apply to the entire table.

While here, fix the one_to_one issue with multiple objects the
same way it was fixed for eager, by picking the first matching
record instead of the last.