Add Database#timezone accessor, for overriding Sequel.database_timezone per Database object
This adds Database#timezone accessor to get and set the database's
timezone, defaulting to Sequel.database_timezone.  It adds
Database#{to,from}_application_timezone, which should now be used
instead of Sequel.database_to_application_timestamp or
Sequel.application_to_database_timestamp.  For these methods to
work without using send, the Sequel.convert_timestamp and
Sequel.convert_output_timestamp methods are now public instead of
private.

This modifies most of Sequel's adapters to use this new timezone
support.  Because conversion procs were previously set per
adapter and now could be modified per-Database because of the
timezone setting, the mysql, sqlite, and ibmdb adapters were
modified to change the conversion procs to be per-Database. This
also allows them to be modified per-Database in general, and you
can use the #conversion_procs accessor to do that.

Because the MySQL convert_tinyint_to_bool and
convert_invalid_date_time settings were relying on the per-adapter
conversion proc hash, they had to be modified to work with the
per-Database conversion proc hash.  So now in addition to the
Sequel::MySQL settings, there are also settings per Database
object, which default to the Sequel::MySQL settings.  This is as
backwards compatible as I could make it, but it still breaks
backwards compatibility in that changing the Sequel::MySQL
setting only affects future Database objects, not already created
ones.