Allow Dataset#insert to return primary key value on Oracle without defining a sequence
This adds Database#autosequence, which can be set to true to have
Dataset#insert guess the sequence value to use for the dataset
based on the table name and Sequel's conventions for defining
sequences.  You can also use the the :autosequence Database option
when creating the Database instance to automatically set the
autosequence value.