Fix handling of booleans in the shared Oracle adapter
Previously, attempts to use boolean columns raised errors.  This
makes attempting to create a boolean column using the generic type
support (TrueClass or FalseClass) to create a char(1) column with
a check for only 'Y' or 'N' values, which seems to be the
recommended way of handling booleans on Oracle, as Oracle has no
native boolean type.  The usage Y/N vs T/F vs 1/0 seems
inconsistent, I choose Y/N since Oracle appears to recommend that.

This changes the literalization of true/false to use 'Y'/'N', but
it's not really possible to return such values as booleans, since
there isn't a native boolean type.  We could add a flag for
assuming that char(1) columns are booleans, but char(1) columns
are much more likely to be used for nonbooleans (unlike tinyint(1)
columns on MySQL or bit(1) types on MSSQL).

Oracle doesn't seem to like WHERE 'Y' for WHERE TRUE (fairly
obvious), and it doesn't seem to support IS TRUE, so set the
appropriate flags on the dataset.