Fix improper handling of emulated bitwise operators with more than 2 arguments
Last month in commit 5722402daa399316a76f09835cc136eb5c6ec389, Sequel
was changed to make the bitwise operators n arity instead of 2 arity,
allowing Sequel to simpler literalization of the bitwise operators,
as it does for the standard mathematical operators:

  :x.sql_number << 1 << 2
  # Before: (("x" << 1) << 2)
  #  After: ("x" << 1 << 2)

Unfortunately, in the cases where the bitwise operators were
emulated, they often assumed that there were exactly 2 arguments,
which caused the third and greater arguments to be ignored.

This adds a Dataset#complex_expression_arg_pairs private method that
yields two arguments at a time to a block, allowing an easy
transition of existing code.  This also changes the adapters that
emulate the bitwise operators to use this new method, and adds
integration tests and specs for the support.