Add Database#listen and #notify to the postgres adapter, for LISTEN and NOTIFY support
Database#notify just does a simple notify, but #listen's
implementation is much more complex.  Because of Sequel's
connection pool, it's hard to be sure that the same connection
object will be used unless you check it out the whole time. You
don't want to start listening on one connection, and wait for
notifications on a different connection.  For that reason, you
can provide a :loop option to #listen that continually waits
for notifications.  This option can also operate as a callback,
allowing you to do processing in between notifications.

Database#listen requires the connection implement wait_for_notify,
so it's currently only added if you are using the postgres adapter
with pg as the underlying driver.  #notify just sends an SQL
string to the database, so it should work on any adapter, so it
is being added to the shared postgres adapter.

This support is currently tested on only PostgreSQL 9.0+, as it
tests the NOTIFY payload option.  However, it should work in
earlier PostgreSQL versions if you don't provide a payload to
notify.