Stop turning off quoting identifiers when running the adapter tests
I'm the one that added this right after 2.2.0 was released, probably
for backwards compatibility with the existing specs.  Unfortunately,
setting this for the adapter specs also sets it for the integration
specs, which can break some the integration specs when they are
with the adapter specs (which is the default the rake tasks use).

This is a fairly large commit as many of the adapter specs are
testing for specific SQL, and almost all of that SQL needed to be
modified to add quotes.