Refactor oracle adapter specs
This cleans them up quite a bit.  I'm removing the testing of for
update, as it causes issues, and we are already testing that the
correct SQL is being used. FOR UPDATE is standard SQL, so we can
assume Oracle implements it properly.

This also has minor changes since identifiers are now quoted by
default in the adapter tests.