Revert the changes to emulate_offset_with_row_number.rb
The attempted optimization of using the ROWNUM psuedocolumn instead
of the ROW_NUMBER window fuction does not work correctly.  This
change means that an order is needed when using an offset on Oracle.
For simplicity, just reverting the changes to
emulate_offset_with_row_number.rb and deleting the overridden
methods in the oracle adapter seem like the easiest approach.