Make Dataset#columns work correctly on jdbc, odbc, ado, and dbi adapters when using an emulated offset on MSSQL and DB2
Before, the shared mssql and db2 adapters modified fetch_rows
to not delete the row number column added by the offset emulation.
However, Dataset#columns was not affected by this.  The correct
approach is to modify the real adapters to remove the columns,
and this modifies the ado, dbi, jdbc, and odbc adapters to do
so.

This adds the private Dataset#offset_returns_row_number_column?
method for seeing if offset adds a row number column, and if
so, it removes it from columns and returned rows. The offset
emulation code sets the method to return true.

This removes the private hash_row method from the dbi adapter.
I didn't actually test the dbi adapter, so hopefully I didn't
break it.