Make Dataset#select_map and #select_order_map return an array of single element arrays if given an array with a single symbol
Due to a programming error on my part, these methods were returning
an array of values instead of wrapping each value in an array.
While here, add a couple of protected methods that the select_map
and select_order_map internal method calls.  This is mostly so they
can be easily overridden later.