Change Dataset#map and #to_hash to not call the row_proc in some cases
This changes #map to not call the row_proc if an argument is given,
and changes #to_hash to not call the row_proc if both arguments are
given.  This makes the methods 1.5-2x faster if called on model
datasets.  It doesn't break any existing specs, but it's possible
that code could be relying on the old behavior.  For model code,
that can happen if your model initialization behavior modifies the
columns in the model object.  However, that is only usually done
for new model objects, not model objects retrieved from the
databases.  For custom row_procs, this could definitely break
code if you are modifying the hash or returning a different object.

This isn't being done for the no-argument form of map, as that is
expected to return model objects for model datasets.  Likewise,
the 1 argument form of to_hash is expected to have model object
values for model datasets.

I don't expect this to break much existing Sequel code, and I
think the performance difference is significant enough to justify
the change.  This also makes #map and #to_hash (and therefore
select_hash) operate more similarly to select_map and
select_order_map, which already removed the row_proc.