Handle CTEs in subselects in more places on databases that don't natively support CTEs in subselects
This adds hoisting of CTEs from datasets given to from, from_self,
with, with_recursive, union, intersect, and except.  Like the
previous support in join_table, it moves the WITH clause from the
subselects into the main query, on datasets that supports CTEs
but not in subselects.

This also fixes the PostgreSQL adapter to work with CTEs in
union/intersect/except by using parantheses around the
SQL, since "UNION WITH cte_name" is a syntax error on PostgreSQL.