Add support for connecting to HSQLDB using the jdbc adapter
This was thankfully fairly easy, as HSQLDB sticks very close to
the SQL standard.  There are a few things that HSQLDB supports
that Sequel does not enable, such as savepoints (HSQLDB seems not
to respect transaction statements in SQL, only the java methods
on the connection object, and Sequel doesn't support savepoints
using java methods in the jdbc transaction support), but almost
everything works (only 2 pending specs).

Note that this will not work with the jdbc-hsqldb gem, as that
uses 1.8.1.3, and the HSQLDB documentation for 1.8.1.3 plain
lies and I don't feel like supporting it (I'll accept patches
that don't break 2.2.5).  You'll need to require the 2.2.5
hsqldb.jar file manually to use this jdbc subadapter, like you
do with the jdbc subadapters that don't have gems, at least
until the jdbc-hsqldb gem is updated.

There are a couple of minor spec changes (other than guards)
that fix specs on HSQLDB.  These changes were made as they
don't affect the purposes of the specs, and the previous
code was causing failures on HSQLDB.