Refactor transaction internals and fix some corner cases
This changes the internal transaction datastructure to be a hash
instead of an array.  Historically, threads were the basis of
the transaction support, so that starting a transaction in
separate threads used separate transactions.  Unfortunately, this
doesn't work if you try to open up two transactions in the same
thread, which can happen if you are using the sharding support.
Also, if you are using a different type of connection pool that
is not based on threads, the transaction code wouldn't have worked.
Additionally, the savepoint level of the current transaction was
stored in a thread local variable.  This doesn't actually cause
problems even if two separate database objects are using savepoints
simultaneously in the same thread, but it's ugly and the fact that
it works is mostly luck.

This changes the basis of the transaction support to be connections
instead of threads, which makes much more fundemental sense, as the
transaction support operates directly with the connection objects.
This allows you to open up multiple transactions simultaneously each
in its own shard in a sharded database configuration.  Since we are
using a hash for data storage, we are now storing the savepoint level
in the hash, so we don't need to use a thread local variable. This
will also make planned future expansion easier.

Unfortunately, this change is fairly invasive, affecting many
adapters.  I should be able to test all of the changed adapters
except the jdbc as400 subadapter, which is changing to use the
native JDBC transaction support, since it included code that was
pretty much identical.

Internally, other than @tranactions changing from an array of threads
to a hash keyed by connection objects, Database#add_transaction now
must accept a connection object.