Remove do (DataObjects) adapter's customized transaction support.
This was the only adapter that used a custom transaction object for
transactions.  It was very clunky, and it wasn't actually used as
all three do subadapters implement savepoints and skipped the
support in favor of their native support.  It's possible that the do
generic support would be more reliable when connecting to something
other than PostgreSQL, MySQL, or SQLite via the do adapter, but
nobody has reported a configuration like that.  I'm going to be
committing changes that will not allow the use of a custom
transaction object, so this has to go now.