Add support for after_commit and after_rollback hooks on Database objects
Using after_commit instead a transaction queues a block that is
called after the commit, and using after_rollback queues a block
that is called after the rollback.

If called outside a transaction, after_commit yields to the block
immediately, and after_rollback ignores the block.  The rationale
here is that the previous changes were autocommitted, so an
after_commit block should be executed and an after_rollback block
should not.

after_commit and after_rollback are not allowed to be called inside
a prepared transaction, as the changes are not guaranteed to be
made when the transaction block returns.  This includes a change so
that the Database object can tell whether an in-progress transaction
on a connection is prepared or not.

This includes more refactoring of the transaction support. This
removes the support for begin_transaction to return a transaction
object to use, which means adapters that override this no longer need
to return the connection object.  This also makes remove_transaction
take a second argument that indicates whether the transaction was
committed.  add_transaction also takes a second argument in order to
see whether the transaction is being prepared or not.

Another change this makes is that if begin_transaction raises an
exception, rollback_transaction is always called.  Before, the
rollback_transaction method wasn't called if begin_transaction
raised an error.  However, there is a race condition there if an
exception occurs after the BEGIN SQL is sent to the database but
before the method returns.  It's safer to always rollback, even
though that may cause another exception.

This requires some changes to adapters that set an autocommit flag on
the underlying connection, as they now need to reset autocommit to
true before calling super in remove_transaction, as otherwise it
would be left false for any after_commit or after_rollback hooks.
Also, they no longer need to check if the conn object is valid,
since they are assured to get passed the underlying connection.

This fixes some variable naming in the commit_or_rollback_transaction
method, where a variable was called thread when it was really a
connection.

This also contains a minor spec change for "should == nil" to
"should be_nil" in some specs