Add after_commit, after_rollback, after_destroy_commit, and after_destroy_rollback hooks to Model objects
These use the recently committed support for after_commit and
after_rollback on Database objects.  after_commit and after_rollback
are called by save, and after_destroy_commit and
after_destroy_rollback by destroy.

Because transactions are optional, after rollback hooks are
called first and after commit hooks are called last.  That way,
if transactions are not being used, after rollback hooks will
be skipped and after commit hooks will be called right before
the method returns.  If transactions are being used,
after_rollback hooks are definitely added, so they will be
called if there is an error, while after_commit hooks are
not added until the end of the method.