Support sharding correctly by default when using transactions in model saving/destroying
Previously, without the sharding plugin, model objects didn't use
the correct shard for transactions.  Model objects should use the
same shard for transactions that they use for inserts, updates,
and deletes.

This pulls the sharding transaction support from the sharding
plugin to the main Sequel code.  It also makes sure that the
after commit and rollback hooks use the correct shard.

This adds a Model#this_server method to get the shard to use for
the model object.  Because Model#this raises exceptions for
models without primary keys, this checks the model's dataset
if no primary key exists.