Add native prepared statement support to the oracle adapter
Unfortunately, ruby-OCI8 doesn't know what to do with nil values
if you don't specify a type, so add in type specifying support
similar to the postgres adapter, using a __* suffix for the
placeholder.

This also catches RuntimeErrors raised when running queries.
ruby-oci8 raises strings, so these should be reraised as
DatabaseErrors.