Work around SQLite column naming bug when using subselects by always aliasing when graphing
This adds Dataset#subselect_columns_require_aliases? private method
for checking whether using referencing a column qualified by a
subselect alias requires a column alias for the correct column
name.  This is false on all databases except SQLite, which includes
the subselect alias qualifier in the column name.

The graphing code has been modified to check this and to always
use aliases if the method returns true (not just with a subselect).
This fixes one of the current integration test failures on SQLite.