Better fix for the SQLite subselect alias qualified column naming bug
The previous fix only fixed the case when you were using graphing,
it didn't fix the general case.  This overrides Dataset#select for
SQLite datasets to check for a subselect and automatically alias
qualified columns if that is the case. It mostly reverts the changes
in 9155f7eadf946bedc6e35121c56aa3feebbaa398, but it keeps the
integration tests and the fix for when the core extensions are not
enabled.  This adds a more general integration test.