Emulate more alter table operations on SQLite
This adds support for emulating add_constraint and drop_constraint
operations.  Sequel duplicates the table if the alter table needs
to be emulated, but it can't completely duplicate everything. For
example, unique constraints are not duplicated, but primary key
and foreign key constraints are.  This makes dropping primary key
constraint drop any primary key constraint.  This makes dropping
any foreign key drop all foreign key constraints on the table,
since the foreign key constraint is by name and Sequel doesn't
know which column is being referenced by the constraint name.

When adding constraints, previous constraints other than primary
key and foreign key constraints are dropped, and the new
constraints are added.

In order for self-referential foreign key constraints to be added
correctly, this changes the order of things when duplicating
a table.  It used to be the new table was created with a backup
name, and then renamed to the correct name.  Now, the old table
is renamed to the backup name, and the new table is created with
the correct name.

In addition to fixing 4 integration tests on SQLite, this also
fixes the handling of composite primary keys when using the
emulated alter table support.