Add Sequel.guarded? to the integration tests
cspecify (short for checked specify) will check this method to
decide whether the block should be pending or not.  This allows
the guarding code to be reused elsewhere, so you don't need to
do INTEGRATION_DB.database_type ==.

This also removes log_specify from the specs, as the plain log
method work fine.