Refactor some association code
Move code for qualifying of association keys with table names into
the association reflection to DRY up quite a bit of code.

Split eager_loading_dataset into two methods so that part of it
can be reused for an upcoming commit.

The :eager_block association option is now called before the
:eager_graph association has been applied instead of after.

A bug when eager loading a many_through_many option with
composite keys that requires an alias for one of the join
tables has been fixed.

Model classes now need to have a dataset before using the
association methods.