Allow Database#schema to take a dataset as the table argument, and use its identifier input/output methods
This is being added so that if you create a model using a dataset
that uses identifier input/output methods different from the
database default, the model will be able to parse the schema
correctly.

To implement this, input_identifier_meth and output_identifier_meth
now take an optional dataset argument, and schema passes the
dataset to schema_parse_table in a :dataset option.  When a model
parses the schema, it will always call Database#schema, as now
Database#schema is responsible for raising an Error if the
dataset it is passed uses more than one table.

All of the adapters that override schema_parse_table need to be
modified for this change, but thankfully the modifications are
fairly easy.

This removes the blind rescue when calling the schema method. Now
Sequel::Errors that are not DatabaseConnectionErrors are the only
errors that will be ignored.