Add mock adapter for better mocking of a database connection
This is mostly intended for use by the specs, but I've also changed
bin/sequel to use it if no database was provided.

Currently, Sequel uses a MockDatabase in the specs, but it has
issues.  It doesn't mock transactions or Sequel's sharding support
correctly, and it doesn't provide any way to set the return values
when using insert, update, or delete, or a way to set the rows
returned by fetch_rows.  Many of the specs that use it do their
own mocking by defining methods, and can be changed to use this new
mock adapter to simplify things.