Initial cleanup of the model specs using the mock adapter
This is the minimum necessary to remove MockDatabase and MockDataset
and replace them with the mock adapter and not have specs break.
Unfortunately, because the model specs shared the same database, it
would have been difficult to do piecemeal.  A more complete
cleanup of the model specs will come later.