Speed up model object instantiation by up to 1-6%
This makes the model dataset row_proc a Method instance instead of
a Proc instance.  I've benchmarked a best case improvement in
performance of 6% on ruby 1.8 and 1% on ruby 1.9.  This certainly
should never be slower, as setting the row_proc directly to the load
method should be more direct than calling the load method in a
regular proc.

Unfortunately, MRI 1.8 is broken for Method objects that call
super which are defined in modules.  To work around this bug, add
a _load method which calls load.  Using this on MRI 1.8 is still
faster than the Proc, since a Proc is a closure and a Method is not.

This does cause an issue if a plugin overrides the load class method,
as then the dataset row proc points to the earlier defined load
method.  To work around that, add a reset_row_proc class method
that plugins that override load must call.