Add Model.call as a replacement for Model.load
This changes the model dataset row_proc to be the model class
itself, since the dataset just calls the call method on the
row_proc.  This fixes the issue on MRI 1.8 where you need a fake
_load method because super in Method arguments is broken, and
it makes it so you don't need to reset the row_proc if you
modify call.  That's the first optimization.

The second optimization is using allocate instead of calling
new with two parameters.  New was just calling set_values and
after_initialize, so call now just does allocate to get a blank
object, then set_values, then after_initialize.  This resulted
in another 3-5% increase in speed, and some more benchmarking
shows a 16% increase in speed on MRI 1.8 compared to 3.28.0, and
a 7% increase in speed on MRI 1.9.

This change should be mostly backwards compatible, and didn't
require a spec change other than an issue caused by too much
mocking.  However, any plugins that used Model.load (either
calling it or overriding it) should switch to Model.call.

Model#set_values is now a public method, since I don't want to
have to use send to call it in Model.call.  I don't encourage its
direct use, though.

This removes the reset_row_proc and _load class methods that were
added recently, they are no longer needed.