Add Database#extend_datasets for the equivalent of extending of the Database object's datasets with a module
I've wanted to something like this for a while, but only previously
considered taking the Sequel::Model approach of using modules for
all methods.  Unfortunately, that approach doesn't work great when
each adapter and subadapter have their own dataset class.

Now that a Database object knows which dataset class to use, we
can make extending the database's datasets simple by subclassing
the current dataset class, setting the database class to the
subclass, and including the module in the subclass.