Make like case sensitive on SQLite and Microsoft SQL Server (use ilike for case insensitive matching)
This makes Sequel on SQLite and Microsoft SQL Server behave more
like other servers, where like is case sensitive and ilike is not.

This adds integration tests to check for correct support of like,
ilike, and their inversions.

For SQLite, this adds support to set the case_sensitive_like
PRAGMA, and sets it to true by default.  For Microsoft SQL
Server, this adds a case sensitive COLLATE clause to both
arguments of LIKE.