Fully support using an aliased table name as the :join_table option for a many_to_many association
Previously, Sequel only had very little support for providing the
:join_table option as an aliased name (e.g. :table___alias).  This
isn't a problem in most code, as about the only time you would need
to do this is using a self-referential association where the join
table is the same as the current or associated table (or both).
The simplest example is a tree structure using a many_to_many to
get all grandchild nodes.

This adds join_table_alias and join_table_source methods to the
many_to_many association reflection.  join_table_alias was already
available via the associated_key_table method, so make that just
an alias for join_table_alias.  join_table_source is the source
without the alias, join_table_alias is the alias, or the source
if there is no alias.

Because of possible join table aliasing, the :qualified_right_key
entry in the many_to_many association reflection now must be
dynamic, so use the qualified_right_key method to get it.

Because of how the eager_grapher proc works, the alias for the
associated table is produced before the alias to the join table,
but the join table is joined to the dataset before the associated
table is joined.  This means that Dataset#unused_table_alias
wouldn't correctly handle the case where you alias the join table
to the same name as the associated table.  Fix this by adding an
optional second argument to unused_table_alias specifying the
default aliases to considered already used.