Fix Model.[] optimization when dataset uses identifier_input_method different than database
Previously, the optimized Model.[] case used the wrong primary key
if the dataset had a different identifier_input_method than the
database.  Also, the Model.[] optimization was not applied if you
gave a dataset to set_dataset.  Now, the optimization is applied
if the dataset is a simple select * from a table.