Switch dataset literalization to be append-only all the way down
Before, dataset literalization was append-only only at the SQL
clause level (e.g. SELECT, FROM, WHERE).  This makes Sequel use
an append-only design all the way down.

This change fixes some bad performance corner cases, and should
result in better performance across the board, with more complex
queries having more benefit.  The worst cases previously occurred
when using large objects (usually strings or blobs), especially
when nested deep in a expression.  This is because those structures
used the following design for literalization:

  "#{literal(subexp1)} #{op} #{literal(subexp2)}"

With deep nesting, every nesting level created its own string which
is at least as large as the level beneath it.  If you are nested 10
levels deep with a large object on the bottom level, you end up
duplicating the large object in memory at least 10 times. This design
passes the sql string being created all the way down, and every level
just appends to that string:

  literal_append(sql, subexp1)
  sql << " #{op} "
  literal_append(sql, subexp2)

This append-only design goes down to the lowest levels, where the
strings themselves are literalized/escaped:

  sql << "'" << escape(string) << "'"

This append-only design performs better at all levels.  It's about
10% better on simple queries, is about twice as fast for any query
using large strings or using a deeply nested structure, and is over
two orders of magnitude faster in some pathological cases (1M string
nested 200 levels deep, 6.25s vs. 1046s).

This design is accomplished by adding *_append methods for all
methods that are used in the literalization process.  These append
methods take the sql string to append to as the first argument,
and just append to that sql string.

This change is backwards compatible in the sense that users that
don't override methods see no change, since all of the *_append
methods still have the regular non-append name, and just call the
*_append method with an empty string, and then return that string
after the *_append method has added to it.

Unfortunately, this change is not backwards compatible if you were
overriding methods and expecting them to get called.  For example,
if you previously overrode Dataset#literal, you should change to
override Dataset#literal_append.

Basically, this should be backwards compatible with applications,
but not backwards compatible with custom plugins, extensions, or
adapters where any method that now has an *_append method was
overridden.

This append-only approach requires that the only operation on the
sql string is appending.  Previously, this wasn't true on databases
that supported common table expressions, as they called replace on
the string.  This was required because previously, the SQL string
creation started with the clause name (e.g. 'SELECT'), and CTEs come
before the clause name.  Now, SQL string creation starts out with a
blank string, and the clause name is added later (after the CTE WITH
clause).  This change will break custom adapters that override
*_clause_methods, those adapters need to be updated to add
the appropriate method to call.

In most cases, this append-only literalization is applied all the way
down.  This even happens with subselects, where they append to the
primary dataset's sql.

There are a few cases that still use the old style in certain
places, mostly when emulating bitwise operators or literalizing
placeholder literal strings.  It may be possible to use an
append-only approach for placeholder literal strings, I'll be
considering that later.

You can get a list of all methods that now have *_append versions by
looking at the Dataset::PUBLIC_APPEND_METHODS and
Dataset::PRIVATE_APPEND_METHODS constants.

In addition to switching to an append-only literalization, there are
additional speedups here for literalizing blobs.  Many adapters used
the following approach:

  blob.each_byte{|x| sql << sprintf('%02x', x)}

However, that approach is very slow compared to:

  sql << blob.unpack("H*").first

This switches all adapters that hex-escape blobs to use unpack.

Other small changes:

* Dataset#on_duplicate_key_update_sql has been removed from the
  shared mysql adapter.
* The :columns dataset option (used when inserting), is no longer
  literalized in advance.  This is used by the insert_columns_sql
  method.
* Symbol splitting is also slightly more performant.
* All Sequel::SQL::Expression objects now respond to to_s_append
  and call the appropriate *_append method on the dataset.
* The sql_literal_append method is now tried first if there is an
  attempt to literalize an object that Sequel does not handle. Sequel
  falls back to trying sql_literal if the sql_literal_append method
  is not defined.
* Dataset#as_sql no longer takes an expression, it only takes the
  alias, and only adds the alias part.
* Many adapters that used the :limit/:offset option verbatim now
  correctly pass it through Dataset#literal.